Clean up publish and edit handler and service logic
Thanks Camila for being a great Go consultant and advising on how to tidy this up!

(I'm still a lil grumpy about some parts, but it's in much better shape 😄)

- Consolidated publisher extension validation into a single function.
- Updated the `Publish` and `EditServer` methods to directly use `XPublisher` from the request.
- Removed outdated tests for publisher extension extraction and validation.
- Introduced new validation functions for server details and remote URL matching.
- Enhanced error handling for server name format and remote URL validation.
- Updated existing tests to reflect changes in validation logic and structure.
