+++ internal/api/handlers/v0/edit.go
@@ -2,7 +2,6 @@ package v0
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"net/http"
 	"strings"
@@ -18,9 +17,9 @@ import (
 
 // EditServerInput represents the input for editing a server
 type EditServerInput struct {
-	Authorization string `header:"Authorization" doc:"Registry JWT token with edit permissions" required:"true"`
-	ID            string `path:"id" doc:"Server ID (UUID)" format:"uuid"`
-	RawBody       []byte `body:"raw"`
+	Authorization string               `header:"Authorization" doc:"Registry JWT token with edit permissions" required:"true"`
+	ID            string               `path:"id" doc:"Server ID (UUID)" format:"uuid"`
+	Body          model.PublishRequest `body:""`
 }
 
 // RegisterEditEndpoints registers the edit endpoint
@@ -53,11 +52,6 @@ func RegisterEditEndpoints(api huma.API, registry service.RegistryService, cfg *
 			return nil, huma.Error401Unauthorized("Invalid or expired Registry JWT token", err)
 		}
 
-		// Validate that only allowed extension fields are present
-		if err := model.ValidatePublishRequestExtensions(input.RawBody); err != nil {
-			return nil, huma.Error400BadRequest("Invalid request format", err)
-		}
-
 		// Get current server to check permissions against existing name
 		currentServer, err := registry.GetByID(input.ID)
 		if err != nil {
@@ -72,30 +66,23 @@ func RegisterEditEndpoints(api huma.API, registry service.RegistryService, cfg *
 			return nil, huma.Error403Forbidden("You do not have edit permissions for this server")
 		}
 
-		// Parse the validated request body
-		var editRequest model.PublishRequest
-		if err := json.Unmarshal(input.RawBody, &editRequest); err != nil {
-			return nil, huma.Error400BadRequest("Invalid JSON format", err)
-		}
-
-		// Validate the server detail
-		validator := validators.NewObjectValidator()
-		if err := validator.Validate(&editRequest.Server); err != nil {
+		// Perform all schema validation
+		if err := validators.ValidatePublishRequest(input.Body); err != nil {
 			return nil, huma.Error400BadRequest(err.Error())
 		}
 
 		// Prevent renaming servers
-		if currentServer.Server.Name != editRequest.Server.Name {
+		if currentServer.Server.Name != input.Body.Server.Name {
 			return nil, huma.Error400BadRequest("Cannot rename server")
 		}
 
 		// Prevent undeleting servers - once deleted, they stay deleted
-		if currentServer.Server.Status == model.ServerStatusDeleted && editRequest.Server.Status != model.ServerStatusDeleted {
+		if currentServer.Server.Status == model.ServerStatusDeleted && input.Body.Server.Status != model.ServerStatusDeleted {
 			return nil, huma.Error400BadRequest("Cannot change status of deleted server. Deleted servers cannot be undeleted.")
 		}
 
 		// Edit the server
-		updatedServer, err := registry.EditServer(input.ID, editRequest)
+		updatedServer, err := registry.EditServer(input.ID, input.Body)
 		if err != nil {
 			if errors.Is(err, database.ErrNotFound) {
 				return nil, huma.Error404NotFound("Server not found")
