+++ internal/api/handlers/v0/publish.go
@@ -2,7 +2,6 @@ package v0
 
 import (
 	"context"
-	"encoding/json"
 	"net/http"
 	"strings"
 
@@ -16,8 +15,8 @@ import (
 
 // PublishServerInput represents the input for publishing a server
 type PublishServerInput struct {
-	Authorization string `header:"Authorization" doc:"Registry JWT token (obtained from /v0/auth/token/github)" required:"true"`
-	RawBody       []byte `body:"raw"`
+	Authorization string               `header:"Authorization" doc:"Registry JWT token (obtained from /v0/auth/token/github)" required:"true"`
+	Body          model.PublishRequest `body:""`
 }
 
 // RegisterPublishEndpoint registers the publish endpoint
@@ -50,33 +49,18 @@ func RegisterPublishEndpoint(api huma.API, registry service.RegistryService, cfg
 			return nil, huma.Error401Unauthorized("Invalid or expired Registry JWT token", err)
 		}
 
-		// Validate that only allowed extension fields are present
-		if err := model.ValidatePublishRequestExtensions(input.RawBody); err != nil {
-			return nil, huma.Error400BadRequest("Invalid request format", err)
-		}
-
-		// Parse the validated request body
-		var publishRequest model.PublishRequest
-		if err := json.Unmarshal(input.RawBody, &publishRequest); err != nil {
-			return nil, huma.Error400BadRequest("Invalid JSON format", err)
-		}
-
-		// Get server details from request body
-		serverDetail := publishRequest.Server
-
-		// Validate the server detail
-		validator := validators.NewObjectValidator()
-		if err := validator.Validate(&serverDetail); err != nil {
+		// Perform all schema validation
+		if err := validators.ValidatePublishRequest(input.Body); err != nil {
 			return nil, huma.Error400BadRequest(err.Error())
 		}
 
-		// Verify that the token's repository matches the server being published
-		if !jwtManager.HasPermission(serverDetail.Name, auth.PermissionActionPublish, claims.Permissions) {
+		// Verify that the token has permission to publish the server
+		if !jwtManager.HasPermission(input.Body.Server.Name, auth.PermissionActionPublish, claims.Permissions) {
 			return nil, huma.Error403Forbidden("You do not have permission to publish this server")
 		}
 
 		// Publish the server with extensions
-		publishedServer, err := registry.Publish(publishRequest)
+		publishedServer, err := registry.Publish(input.Body)
 		if err != nil {
 			return nil, huma.Error400BadRequest("Failed to publish server", err)
 		}
