+++ internal/api/handlers/v0/publish_integration_test.go
@@ -168,7 +168,9 @@ func TestPublishIntegration(t *testing.T) {
 	t.Run("publish fails with invalid token", func(t *testing.T) {
 		publishReq := model.PublishRequest{
 			Server: model.ServerDetail{
-				Name: "test-server",
+				Name:        "io.github.domdomegg/test-server",
+				Description: "Test server",
+				VersionDetail: model.VersionDetail{Version: "1.0.0"},
 			},
 		}
