+++ internal/database/import.go
@@ -15,7 +15,7 @@ import (
 
 // ReadSeedFile reads seed data from various sources:
 // 1. Local file paths (*.json files) - expects extension wrapper format
-// 2. Direct HTTP URLs to seed.json files - expects extension wrapper format  
+// 2. Direct HTTP URLs to seed.json files - expects extension wrapper format
 // 3. Registry root URLs (automatically appends /v0/servers and paginates)
 // Only the extension wrapper format is supported (array of ServerResponse objects)
 func ReadSeedFile(ctx context.Context, path string) ([]*model.ServerRecord, error) {
@@ -127,7 +127,7 @@ func fetchFromRegistryAPI(ctx context.Context, baseURL string) ([]*model.ServerR
 func convertServerResponseToRecord(response model.ServerResponse) *model.ServerRecord {
 	// Extract registry metadata from the extension
 	registryExt := response.XIOModelContextProtocolRegistry
-	
+
 	// Parse timestamps
 	publishedAt, _ := time.Parse(time.RFC3339, getStringFromInterface(registryExt, "published_at"))
 	updatedAt, _ := time.Parse(time.RFC3339, getStringFromInterface(registryExt, "updated_at"))
@@ -141,11 +141,9 @@ func convertServerResponseToRecord(response model.ServerResponse) *model.ServerR
 	}
 
 	// Publisher extensions
-	publisherExtensions := make(map[string]interface{})
-	if response.XPublisher != nil {
-		if publisherMap, ok := response.XPublisher.(map[string]interface{}); ok {
-			publisherExtensions = publisherMap
-		}
+	publisherExtensions := response.XPublisher
+	if publisherExtensions == nil {
+		publisherExtensions = make(map[string]interface{})
 	}
 
 	return &model.ServerRecord{
