+++ internal/service/registry_service.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/url"
-	"strings"
 	"time"
 
 	"github.com/google/uuid"
@@ -70,159 +68,61 @@ func (s *registryServiceImpl) GetByID(id string) (*model.ServerResponse, error) 
 	return &response, nil
 }
 
-// validateMCPBPackage validates MCPB packages
-func validateMCPBPackage(host string) error {
-	allowedHosts := []string{
-		"github.com",
-		"www.github.com",
-		"gitlab.com",
-		"www.gitlab.com",
-	}
-
-	isAllowed := false
-	for _, allowed := range allowedHosts {
-		if host == allowed {
-			isAllowed = true
-			break
-		}
-	}
-
-	if !isAllowed {
-		return fmt.Errorf("MCPB packages must be hosted on allowlisted providers (GitHub or GitLab). Host '%s' is not allowed", host)
-	}
-
-	return nil
-}
-
-// validatePackage validates packages to ensure they meet requirements
-func validatePackage(pkg *model.Package) error {
-	registryType := strings.ToLower(pkg.RegistryType)
-
-	// For direct download packages (mcpb or direct URLs)
-	if registryType == "mcpb" ||
-		strings.HasPrefix(pkg.Identifier, "http://") || strings.HasPrefix(pkg.Identifier, "https://") {
-		parsedURL, err := url.Parse(pkg.Identifier)
-		if err != nil {
-			return fmt.Errorf("invalid package URL: %w", err)
-		}
-
-		host := strings.ToLower(parsedURL.Host)
-
-		// For MCPB packages, validate they're from allowed hosts
-		if registryType == "mcpb" || strings.HasSuffix(strings.ToLower(pkg.Identifier), ".mcpb") {
-			return validateMCPBPackage(host)
-		}
-
-		// For other URL-based packages, just ensure it's valid
-		if parsedURL.Scheme == "" || parsedURL.Host == "" {
-			return fmt.Errorf("package URL must be a valid absolute URL")
-		}
-		return nil
-	}
-
-	// For registry-based packages, no special validation needed
-	// Registry types like "npm", "pypi", "docker-hub", "nuget" are all valid
-	return nil
-}
-
 // Publish publishes a server with separated extensions
 func (s *registryServiceImpl) Publish(req model.PublishRequest) (*model.ServerResponse, error) {
 	// Create a timeout context for the database operation
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 
-	// Validate the request
-	if err := model.ValidatePublisherExtensions(req); err != nil {
-		return nil, err
-	}
-
-	// Validate server name exists and format
-	if _, err := model.ParseServerName(req.Server); err != nil {
-		return nil, err
-	}
-
-	// Validate all packages
-	for _, pkg := range req.Server.Packages {
-		if err := validatePackage(&pkg); err != nil {
-			return nil, fmt.Errorf("validation failed: %w", err)
-		}
-	}
-
-	// Validate reverse-DNS namespace matching for remote URLs
-	if err := model.ValidateRemoteNamespaceMatch(req.Server); err != nil {
-		return nil, err
-	}
+	publishTime := time.Now()
 
 	// Check for duplicate remote URLs
 	if err := s.validateNoDuplicateRemoteURLs(ctx, req.Server); err != nil {
 		return nil, err
 	}
 
-	// Get the new version's details
-	newVersion := req.Server.VersionDetail.Version
-	newName := req.Server.Name
-	currentTime := time.Now()
-
-	// Check for existing versions of this server
-	existingServers, _, err := s.db.List(ctx, map[string]any{"name": newName}, "", maxServerVersionsPerServer)
+	existingServerVersions, _, err := s.db.List(ctx, map[string]any{"name": req.Server.Name}, "", maxServerVersionsPerServer)
 	if err != nil && !errors.Is(err, database.ErrNotFound) {
 		return nil, err
 	}
 
-	if len(existingServers) == maxServerVersionsPerServer {
+	// Check we haven't exceeded the maximum versions allowed for a server
+	if len(existingServerVersions) >= maxServerVersionsPerServer {
 		return nil, database.ErrMaxServersReached
 	}
 
-	// Determine if this version should be marked as latest
-	isLatest := true
-	var existingLatestID string
-
-	// Check all existing versions for duplicates and determine if new version should be latest
-	for _, server := range existingServers {
+	// Check this isn't a duplicate version
+	for _, server := range existingServerVersions {
 		existingVersion := server.ServerJSON.VersionDetail.Version
-
-		// Early exit: check for duplicate version
-		if existingVersion == newVersion {
+		if existingVersion == req.Server.VersionDetail.Version {
 			return nil, database.ErrInvalidVersion
-		}
-
-		if server.RegistryMetadata.IsLatest {
-			existingLatestID = server.RegistryMetadata.ID
-			existingTime, _ := time.Parse(time.RFC3339, server.RegistryMetadata.ReleaseDate)
-
-			// Compare versions using the proper versioning strategy
-			comparison := CompareVersions(newVersion, existingVersion, currentTime, existingTime)
-			if comparison <= 0 {
-				// New version is not greater than existing latest
-				isLatest = false
-			}
 		}
 	}
 
-	// Extract publisher extensions from request
-	publisherExtensions := model.ExtractPublisherExtensions(req)
+	// Determine if this version should be marked as latest
+	existingLatest := s.getCurrentLatestVersion(existingServerVersions)
+	isNewLatest := CompareVersions(req.Server.VersionDetail.Version, existingLatest.ServerJSON.VersionDetail.Version, publishTime, existingLatest.RegistryMetadata.PublishedAt) > 0
 
 	// Create registry metadata with service-determined values
 	registryMetadata := model.RegistryMetadata{
 		ID:          uuid.New().String(),
-		PublishedAt: currentTime,
-		UpdatedAt:   currentTime,
-		IsLatest:    isLatest,
-		ReleaseDate: currentTime.Format(time.RFC3339),
-	}
-
-	// If this will be the latest version, we need to update the existing latest
-	if isLatest && existingLatestID != "" {
-		// Update the existing latest to no longer be latest
-		if err := s.db.UpdateLatestFlag(ctx, existingLatestID, false); err != nil {
-			return nil, err
-		}
+		PublishedAt: publishTime,
+		UpdatedAt:   publishTime,
+		IsLatest:    isNewLatest,
+		ReleaseDate: publishTime.Format(time.RFC3339),
 	}
 
 	// Publish to database with the registry metadata
-	serverRecord, err := s.db.Publish(ctx, req.Server, publisherExtensions, registryMetadata)
+	serverRecord, err := s.db.Publish(ctx, req.Server, req.XPublisher, registryMetadata)
 	if err != nil {
 		return nil, err
+	}
+
+	// Mark previous latest as no longer latest
+	if isNewLatest && existingLatest != nil {
+		if err := s.db.UpdateLatestFlag(ctx, existingLatest.RegistryMetadata.ID, false); err != nil {
+			return nil, err
+		}
 	}
 
 	// Convert ServerRecord to ServerResponse format
@@ -255,40 +155,28 @@ func (s *registryServiceImpl) validateNoDuplicateRemoteURLs(ctx context.Context,
 	return nil
 }
 
+// getCurrentLatestVersion finds the current latest version from existing server versions
+func (s *registryServiceImpl) getCurrentLatestVersion(existingServerVersions []*model.ServerRecord) *model.ServerRecord {
+	for _, server := range existingServerVersions {
+		if server.RegistryMetadata.IsLatest {
+			return server
+		}
+	}
+	return nil
+}
+
 // EditServer updates an existing server with new details (admin operation)
 func (s *registryServiceImpl) EditServer(id string, req model.PublishRequest) (*model.ServerResponse, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
-
-	// Validate the request
-	if err := model.ValidatePublisherExtensions(req); err != nil {
-		return nil, err
-	}
-
-	// Validate server name exists and format
-	if _, err := model.ParseServerName(req.Server); err != nil {
-		return nil, err
-	}
-
-	// Validate all packages
-	for _, pkg := range req.Server.Packages {
-		if err := validatePackage(&pkg); err != nil {
-			return nil, fmt.Errorf("validation failed: %w", err)
-		}
-	}
-
-	// Validate reverse-DNS namespace matching for remote URLs
-	if err := model.ValidateRemoteNamespaceMatch(req.Server); err != nil {
-		return nil, err
-	}
 
 	// Check for duplicate remote URLs
 	if err := s.validateNoDuplicateRemoteURLs(ctx, req.Server); err != nil {
 		return nil, err
 	}
 
-	// Extract publisher extensions from request
-	publisherExtensions := model.ExtractPublisherExtensions(req)
+	// Use publisher extensions directly from request
+	publisherExtensions := req.XPublisher
 
 	// Update server in database
 	serverRecord, err := s.db.UpdateServer(ctx, id, req.Server, publisherExtensions)
